requirements.txt
# Web Framework
Flask==2.3.2
gunicorn==21.2.0

# Browser Automation
selenium==4.10.0
webdriver-manager==3.8.6

# PDF Processing
pdfplumber==0.9.0
PyMuPDF==1.23.0

# HTTP Requests
requests==2.31.0

# Date/Time Handling
python-dateutil==2.8.2
pytz==2023.3

# Configuration
PyYAML==6.0.1

# Database (if using SQLite extensions)
# Add any other packages your code imports

# Development/Testing (optional)
# pytest==7.4.0
# black==23.3.0
